+++ CHANGELOG.md
@@ -29,6 +29,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Ignore empty file names (e.g. `""`) rather than emitting a warning
 - Omit superfluous prb/drv for trivial signal connections in instances (#205)
 - Fix modport selection on interface instances (#199)
+- Fix wire declarations requiring default assignment to be constant (#198)
 
 ## 0.10.0 - 2020-06-15
 ### Added